chore: ENABLING-204, configure vitest on client and react packages
