+++ package.json
@@ -33,10 +33,12 @@
     "preinstall": "npx only-allow pnpm",
     "lint": "turbo run lint --filter=./packages/**",
     "pre-commit": "pnpm run lint && pnpm run format",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "test": "turbo run test --filter=./packages/react"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^3.2.0",
+    "@eslint/js": "catalog:",
     "@skypack/package-check": "^0.2.2",
     "@storybook/addon-a11y": "^8.3.6",
     "@storybook/addon-essentials": "^8.3.6",
@@ -48,19 +50,23 @@
     "@storybook/react-vite": "^8.3.6",
     "@storybook/test": "^8.3.6",
     "@storybook/theming": "^8.3.6",
+    "@testing-library/jest-dom": "6.5.0",
+    "@testing-library/react": "16.0.1",
+    "@testing-library/user-event": "14.5.2",
     "@turbo/gen": "^2.3.1",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@typescript-eslint/parser": "^6.15.0",
+    "@vitest/coverage-v8": "^2.1.3",
+    "@vitest/ui": "^2.1.3",
     "chromatic": "^11.18.1",
-    "@eslint/js": "catalog:",
     "eslint": "catalog:",
-    "typescript-eslint": "catalog:",
-    "globals": "^15.11.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-storybook": "^0.10.1",
+    "globals": "^15.11.0",
     "husky": "^9.1.7",
+    "jsdom": "^25.0.1",
     "msw": "^2.3.1",
     "msw-storybook-addon": "^2.0.3",
     "prettier": "catalog:",
@@ -68,7 +74,9 @@
     "rollup-plugin-visualizer": "5.12.0",
     "storybook": "^8.3.6",
     "turbo": "catalog:",
-    "typescript": "catalog:"
+    "typescript": "catalog:",
+    "typescript-eslint": "catalog:",
+    "vitest": "^2.1.3"
   },
   "packageManager": "pnpm@9.12.2",
   "engines": {
