+++ packages/bootstrap/package.json
@@ -28,11 +28,11 @@
     "clean": "rm -rf dist",
     "compile": "sass --load-path=node_modules/ --style=compressed src/index.scss dist/index.css",
     "fix": "stylelint src/**/*.scss --fix",
-    "lint": "stylelint src/**/*.scss",
     "format": "pnpm format:write && pnpm format:check",
     "format:check": "npx prettier --check src/**/*.scss",
     "format:write": "npx prettier --write src/**/*.scss",
     "generate-commit-version": "node scripts/version.cjs",
+    "lint": "stylelint src/**/*.scss",
     "watch": "sass --watch --load-path=node_modules/ --style=compressed src/index.scss dist/index.css"
   },
   "dependencies": {
