+++ packages/client/package.json
@@ -40,7 +40,8 @@
     "format": "pnpm run format:write && pnpm run format:check",
     "format:check": "npx prettier --check \"src/ts/**/*.ts\"",
     "format:write": "npx prettier --write \"src/ts/**/*.ts\"",
-    "lint": "eslint"
+    "lint": "eslint",
+    "test": "vitest"
   },
   "dependencies": {
     "@edifice.io/utilities": "workspace:*",
@@ -48,6 +49,7 @@
     "core-js": "^3.35.1"
   },
   "devDependencies": {
+    "@edifice.io/config": "workspace:*",
     "@types/jasmine": "5.1.4",
     "@types/node": "catalog:",
     "jasmine": "5.1.0",
