--- packages/client/src/test/config/jasmine.json
@@ -1,11 +1,0 @@
-{
-  "spec_dir": "./test",
-  "spec_files": [
-    "**/*[sS]pec.?(m)js"
-  ],
-  "helpers": [
-    "helpers/**/*.js"
-  ],
-  "stopSpecOnExpectationFailure": false,
-  "random": false
-}
