--- packages/client/src/test/mocks/data/MockedContextData.json
@@ -1,34 +1,0 @@
-{
-  "folders":[]
- ,"filters":[]
- ,"orders":[{
-     "id": "name"
-    ,"defaultValue": "asc"
-    ,"i18n": "fake.key.order.name.asc"
-  }
- ]
- ,"actions":[
-      {"id":"comment",  "available":false, "workflow": ""}
-     ,{"id":"copy",     "available":false, "workflow": ""}
-     ,{"id":"create",   "available":false, "workflow": ""}
-     ,{"id":"delete",   "available":false, "workflow": ""}
-     ,{"id":"export",   "available":false, "workflow": ""}
-     ,{"id":"initialize", "available":true, "workflow": ""}
-     ,{"id":"manage",   "available":false, "workflow": ""}
-     ,{"id":"move",     "available":false, "workflow": ""}
-     ,{"id":"open",     "available":false, "workflow": ""}
-     ,{"id":"print",    "available":false, "workflow": ""}
-     ,{"id":"publish",  "available":false, "workflow": ""}
-     ,{"id":"search",   "available":true, "workflow": ""}
-     ,{"id":"share",    "available":false, "workflow": ""}
-    ]
- ,"pagination":{
-     "startIdx": 0
-    ,"maxIdx": 22
-    ,"pageSize": 10
- }
- ,"resources":[]
- ,"preferences":{
-     "view": "list"
- }
-}
