--- packages/client/src/test/mocks/data/apps-sharing-rights.json
@@ -1,35 +1,0 @@
-{
-"blog": {
-    "blog.read": [
-        "org-entcore-blog-controllers-PostController|list",
-        "org-entcore-blog-controllers-BlogController|get",
-        "org-entcore-blog-controllers-PostController|comments",
-        "org-entcore-blog-controllers-PostController|get"
-    ],
-    "blog.manager": [
-        "org-entcore-blog-controllers-BlogController|update",
-        "org-entcore-blog-controllers-BlogController|delete",
-        "org-entcore-blog-controllers-PostController|publishComment",
-        "org-entcore-blog-controllers-BlogController|shareResource",
-        "org-entcore-blog-controllers-BlogController|shareJsonSubmit",
-        "org-entcore-blog-controllers-BlogController|shareJson",
-        "org-entcore-blog-controllers-PostController|publish",
-        "org-entcore-blog-controllers-BlogController|updatePublicBlog",
-        "org-entcore-blog-controllers-BlogController|removeShare"
-    ],
-    "blog.contrib": [
-        "org-entcore-blog-controllers-PostController|unpublish",
-        "org-entcore-blog-controllers-PostController|create",
-        "org-entcore-blog-controllers-PostController|delete",
-        "org-entcore-blog-controllers-PostController|submit",
-        "org-entcore-blog-controllers-PostController|update"
-    ],
-    "blog.comment": [
-        "org-entcore-blog-controllers-PostController|updateComment",
-        "org-entcore-blog-controllers-PostController|deleteComment",
-        "org-entcore-blog-controllers-PostController|comment"
-    ]
-}
-,"exercizer":{"exercizer.manager":["fr-openent-exercizer-controllers-GrainScheduledController|persist","fr-openent-exercizer-controllers-SubjectController|shareSubmit","fr-openent-exercizer-controllers-SubjectController|remove","fr-openent-exercizer-controllers-SubjectController|shareResource","fr-openent-exercizer-controllers-SubjectController|shareRemove"],"exercizer.contrib":["fr-openent-exercizer-controllers-SubjectController|duplicateGrainsIntoSubject","fr-openent-exercizer-controllers-SubjectScheduledController|persist","fr-openent-exercizer-controllers-SubjectController|update","fr-openent-exercizer-controllers-SubjectController|grainPersist","fr-openent-exercizer-controllers-SubjectController|grainRemoves","fr-openent-exercizer-controllers-SubjectScheduledController|schedule","fr-openent-exercizer-controllers-SubjectScheduledController|simpleSchedule","fr-openent-exercizer-controllers-SubjectController|grainUpdate","fr-openent-exercizer-controllers-SubjectController|canSchedule"],"exercizer.read":["fr-openent-exercizer-controllers-SubjectController|grainList","fr-openent-exercizer-controllers-SubjectController|getMoodle","fr-openent-exercizer-controllers-SubjectController|duplicateSubjects"]}
-,"folder":{}
-}
