--- packages/client/src/test/mocks/data/sharing-rights.json
@@ -1,38 +1,0 @@
-{
-    "read": {
-        "priority": 0,
-        "default": true,
-        "requires": []
-    },
-    "contrib": {
-        "priority": 1,
-        "default": false,
-        "requires": [
-            "read"
-        ]
-    },
-    "publish": {
-        "priority": 2,
-        "default": false,
-        "requires": [
-            "read",
-            "contrib"
-        ]
-    },
-    "manager": {
-        "priority": 3,
-        "default": false,
-        "requires": [
-            "read",
-            "contrib",
-            "publish"
-        ]
-    },
-    "comment": {
-        "priority": 4,
-        "default": true,
-        "requires": [
-            "read"
-        ]
-    }
-}
