+++ packages/client/vite.config.ts
@@ -1,3 +1,5 @@
+/// <reference types="vitest/config" />
+
 import { resolve } from 'path';
 import { visualizer } from 'rollup-plugin-visualizer';
 import { defineConfig, PluginOption } from 'vite';
@@ -18,4 +20,13 @@ export default defineConfig({
     },
   },
   plugins: [dts(), visualizer() as PluginOption],
+
+  test: {
+    watch: false,
+    globals: true,
+    environment: 'jsdom',
+    include: ['src/**/*.spec.ts'],
+    setupFiles: ['./vitest.setup.ts'],
+    reporters: ['default'],
+  },
 });
