+++ packages/config/src/test/index.ts
@@ -1,0 +1,2 @@
+export * from './msw/handlers';
+export * from './msw/mocks';
