+++ packages/config/src/test/msw/handlers.ts
@@ -1,0 +1,23 @@
+import {
+  authHandlers,
+  blogHandlers,
+  commonHandlers,
+  directoryHandlers,
+  publicConfigHandlers,
+  themeHandlers,
+  userbookHandlers,
+  wikiHandlers,
+  workspaceHandlers,
+} from './mocks';
+
+export const handlers = [
+  ...userbookHandlers,
+  ...workspaceHandlers,
+  ...authHandlers,
+  ...blogHandlers,
+  ...commonHandlers,
+  ...directoryHandlers,
+  ...themeHandlers,
+  ...wikiHandlers,
+  ...publicConfigHandlers,
+];
