+++ packages/config/src/test/msw/index.ts
@@ -1,0 +1,2 @@
+export * from './handlers';
+export * from './mocks';
