+++ packages/config/src/test/msw/mocks/blog.ts
@@ -1,0 +1,7 @@
+import { http, HttpResponse } from 'msw';
+
+export const handlers = [
+  http.get('/blog/linker', () => {
+    return HttpResponse.json(null);
+  }),
+];
