+++ packages/config/src/test/msw/mocks/index.ts
@@ -1,0 +1,9 @@
+export { handlers as authHandlers } from './auth';
+export { handlers as blogHandlers } from './blog';
+export { handlers as commonHandlers } from './common';
+export { handlers as publicConfigHandlers } from './config';
+export { handlers as directoryHandlers } from './directory';
+export { handlers as themeHandlers } from './theme';
+export { handlers as userbookHandlers } from './userbook';
+export { handlers as wikiHandlers } from './wiki';
+export { handlers as workspaceHandlers } from './workspace';
