+++ packages/config/tsconfig.json
@@ -3,7 +3,7 @@
     "target": "ESNext",
     "useDefineForClassFields": true,
     "lib": ["ESNext", "DOM"],
-    "moduleResolution": "Node",
+    "moduleResolution": "bundler",
     "strict": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
@@ -13,8 +13,7 @@
     "noUnusedParameters": false,
     "noImplicitReturns": false,
     "skipLibCheck": true,
-    "sourceMap": true,
-    "types": ["vite/client", "vitest/globals", "vitest/importMeta", "vitest"]
+    "sourceMap": true
   },
   "include": ["src"],
   "exclude": ["node_modules"]
