+++ packages/react/src/providers/index.ts
@@ -4,3 +4,4 @@ export * from './EdificeClientProvider/EdificeClientProvider.hook';
 export * from './EdificeThemeProvider/EdificeThemeProvider';
 export * from './EdificeThemeProvider/EdificeThemeProvider.context';
 export * from './EdificeThemeProvider/EdificeThemeProvider.hook';
+export * from './MockedProvider/MockedProvider';
