+++ packages/react/tsconfig.lib.json
@@ -19,7 +19,17 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "types": ["node", "vite/client"]
+    "types": [
+      "node",
+      "vite/client",
+      "vitest/globals",
+      "vitest/importMeta",
+      "vitest"
+    ],
+    "baseUrl": ".",
+    "paths": {
+      "~/setup": ["vitest.setup.ts"]
+    }
   },
   "include": ["src"]
 }
