+++ pnpm-workspace.yaml
@@ -27,3 +27,4 @@ catalog:
   'eslint-plugin-react-hooks': 5.1.0-rc.0
   'eslint-plugin-react-refresh': ^0.4.12
   'typescript-eslint': ^8.8.1
+  'msw': ^2.3.1
