+++ turbo.json
@@ -12,6 +12,9 @@
     "dev": {
       "cache": false
     },
+    "test": {
+      "cache": false
+    },
     "docs": {
       "cache": false
     },
